Allow resource without path when falling back to the root well-known PRM document
This is stricter than the TypeScript SDK which only checks that the resource prefix matches:
https://github.com/modelcontextprotocol/typescript-sdk/blob/06a4fd2332cd0ba8884e18b21ef4f7d03dea7b0d/src/shared/auth-utils.ts#L25

However RFC 9728 makes it clear that the resource name must be identical.

The resource value returned MUST be identical to the protected resource's resource identifier value into which the well-known URI path suffix was inserted to create the URL used to retrieve the metadata. If these values are not identical, the data contained in the response MUST NOT be used.

If the protected resource metadata was retrieved from a URL returned by the protected resource via the WWW-Authenticate resource_metadata parameter, then the resource value returned MUST be identical to the URL that the client used to make the request to the resource server. If these values are not identical, the data contained in the response MUST NOT be used.

https://datatracker.ietf.org/doc/html/rfc9728/#section-3.3